+++ tests/test_devices.py
@@ -1961,6 +1961,38 @@ def test_motion_detector_switch_outdoor(fake_home: Home):
     with no_ssl_verification():
         d = fake_home.search_device_by_id("3014F711000000000SMO230A")
         assert isinstance(d, MotionDetectorSwitchOutdoor)
+        assert d.sabotage is False
+        assert d.motionDetected is False
+        assert d.illumination == 3158.0
+        assert d.currentIllumination is None
+        assert d.motionBufferActive is True
+        assert d.motionDetectionSendInterval == MotionDetectionSendInterval.SECONDS_30
+        assert d.numberOfBrightnessMeasurements == 3
+        assert d.on is False
+        assert d.profileMode == "AUTOMATIC"
+        assert d.userDesiredProfileMode == "AUTOMATIC"
+        assert "motionDetected(False)" in str(d)
+        assert "illumination(3158.0)" in str(d)
+        assert "on(False)" in str(d)
+        assert "sabotage(False)" in str(d)
+
+        motion_channel = d.get_functional_channel(FunctionalChannelType.MOTION_DETECTION_CHANNEL)
+        switch_channel = d.get_functional_channel(FunctionalChannelType.SWITCH_CHANNEL)
+        sabotage_channel = d.get_functional_channel(FunctionalChannelType.DEVICE_SABOTAGE)
+
+        assert isinstance(motion_channel, MotionDetectionChannel)
+        assert isinstance(switch_channel, SwitchChannel)
+        assert isinstance(sabotage_channel, DeviceSabotageChannel)
+
+        assert motion_channel.motionDetected is False
+        assert motion_channel.illumination == 3158.0
+        assert motion_channel.motionBufferActive is True
+        assert motion_channel.motionDetectionSendInterval == MotionDetectionSendInterval.SECONDS_30
+        assert motion_channel.numberOfBrightnessMeasurements == 3
+        assert switch_channel.on is False
+        assert switch_channel.profileMode == "AUTOMATIC"
+        assert switch_channel.userDesiredProfileMode == "AUTOMATIC"
+        assert sabotage_channel.sabotage is False
 
 
 def test_wall_mounted_keypad(fake_home: Home):